add correct import and unify the test cases
